# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Joe Kuntani Platform API
  version: 0.1.0
  description: Versioned contract for the standalone Joe Kuntani platform.
  license:
    name: Proprietary - all rights reserved
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Relative API base; clients configure the environment origin.
security: []
...
paths:
...
  /api/public/newsletter:
    post:
      operationId: subscribeNewsletter
      summary: Subscribe an email to the public newsletter with consent
      tags: [Newsletter]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, consent, consent_version]
              properties:
                email: { type: string, format: email }
                name: { type: string, maxLength: 120 }
                source: { type: string, maxLength: 80 }
                consent: { type: boolean }
                consent_version: { type: string, minLength: 1, maxLength: 40 }
      responses:
        "201":
          description: Subscriber created or reactivated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NewsletterSubscriber" }
        "422": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /api/admin/newsletter:
    get:
      operationId: listNewsletterSubscribers
      summary: List newsletter subscribers
      tags: [Newsletter]
      security: [{ staffSession: [] }]
      responses:
        "200":
          description: Subscribers
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [subscribers]
                properties:
                  subscribers:
                    type: array
                    items: { $ref: "#/components/schemas/NewsletterSubscriber" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
  /api/admin/newsletter/{id}/unsubscribe:
    post:
      operationId: unsubscribeNewsletterSubscriber
      summary: Mark a newsletter subscriber as unsubscribed
      tags: [Newsletter]
      security: [{ staffSession: [], csrfToken: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Unsubscribed }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    staffSession:
      type: apiKey
      in: cookie
      name: jk_admin_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  responses:
    Problem:
      description: Request rejected
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    NewsletterSubscriber:
      type: object
      additionalProperties: false
      required:
        [id, email, source, status, consent_version, consent_at, created_at]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
        source: { type: string }
        status: { type: string }
        consent_version: { type: string }
        consent_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        unsubscribed_at:
          type: [string, "null"]
          format: date-time
...
    Problem:
      type: object
      additionalProperties: true
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference